backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
python-multipart
pydantic
pgvector
psycopg2
requests
google-generativeai
psycopg2-binary
langdetect
googletrans==4.0.0-rc1
google-cloud-speech
google-cloud-texttospeech
httpx
ultralytics
pillow
passlib
python-jose